requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.9
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.16.0
docker==7.1.0
filelock==3.17.0
fonttools==4.56.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.3
packaging==24.2
pdfkit==1.0.0
pillow==11.1.0
platformdirs==4.3.6
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pywin32==310
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.29.3
